fix transaction dashboard

+++ gj/src/main/java/com/example/gj/viewmodel/dash_board/GetTransactionDashBoardResponse.java
@@ -11,5 +11,5 @@ public class GetTransactionDashBoardResponse {
     private TransactionDashBoardResponse totalTransaction;
     private List<TransactionDashBoardResponse> listTransaction;
     private long count;
-    private long sumCost;
+    private double sumCost;
 }
